refactor: restructure renderer ipc code
