+++ apps/electron/public/game/index.html
@@ -7,6 +7,7 @@
 	require('../../dist/renderer/game/api/Bot.js');
 	require('../../dist/renderer/game/handlers/index.js');
 	require('../../dist/renderer/game/script.js');
+	require('../../dist/renderer/game/flash-ipc.js');
 </script>
 
 <div id="gameContainer">
