+++ apps/electron/src/common/ipc-events.ts
@@ -1,0 +1,7 @@
+export const IPC_EVENTS = {
+	GET_DOCUMENTS_PATH: 'root:get-documents-path' as const,
+	TOGGLE_DEV_TOOLS: 'root:toggle-dev-tools' as const,
+	LOGIN: 'root:login' as const,
+	LOGIN_SUCCESS: 'root:login-success' as const,
+	LOAD_SCRIPT: 'root:load_script' as const,
+} as const;
