--- apps/electron/src/common/ipc.events.ts
@@ -1,4 +1,0 @@
-export const IPC_EVENTS = {
-	MAIN: {},
-	RENDERER: {},
-} as const;
